refactor: rename page 'code '
